pyproject.toml
[tool.poetry]
name = "django-azure-auth-testbench"
version = "0.1.0"
description = ""
authors = ["Armand Rego <armand@weirdsheeplabs.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.4"
django-azure-auth = { git = "https://github.com/Weird-Sheep-Labs/django-azure-auth", branch = "main" }
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.2"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
